unify middleware validation
